+++ NRedberry.Core/Combinatorics/Permutation.cs
@@ -4,23 +4,23 @@ using NRedberry.Core.Combinatorics;
 
 public class Permutation : IComparable<Permutation>
 {
-    protected int[] permutation;
+    protected long[] permutation;
 
     public Permutation(int dimension)
     {
-        permutation = new int[dimension];
+        permutation = new long[dimension];
         for (int i = 0; i < dimension; ++i)
             permutation[i] = i;
     }
 
-    public Permutation(int[] permutation)
+    public Permutation(long[] permutation)
     {
         if (!Combinatorics.TestPermutationCorrectness(permutation))
             throw new ArgumentException("Wrong permutation input: input array is not consistent with one-line notation");
-        this.permutation = (int[])permutation.Clone();
+        this.permutation = (long[])permutation.Clone();
     }
 
-    protected Permutation(int[] permutation, bool notClone)
+    protected Permutation(long[] permutation, bool notClone)
     {
         this.permutation = permutation;
     }
@@ -30,11 +30,11 @@ public class Permutation : IComparable<Permutation>
         return new Permutation(permutation.Length);
     }
 
-    protected int[] CompositionArray(Permutation element)
+    protected long[] CompositionArray(Permutation element)
     {
         if (permutation.Length != element.permutation.Length)
             throw new ArgumentException("different dimensions of compositing combinatorics");
-        int[] perm = new int[permutation.Length];
+        long[] perm = new long[permutation.Length];
         for (int i = 0; i < permutation.Length; ++i)
             perm[i] = element.permutation[permutation[i]];
         return perm;
@@ -55,15 +55,15 @@ public class Permutation : IComparable<Permutation>
         return copy;
     }
 
-    protected int[] CalculateInverse()
+    protected long[] CalculateInverse()
     {
-        int[] inverse = new int[permutation.Length];
-        for (int i = 0; i < permutation.Length; ++i)
+        long[] inverse = new long[permutation.Length];
+        for (long i = 0; i < permutation.Length; ++i)
             inverse[permutation[i]] = i;
         return inverse;
     }
 
-    public int NewIndexOf(int index)
+    public long NewIndexOf(long index)
     {
         return permutation[index];
     }
@@ -73,7 +73,7 @@ public class Permutation : IComparable<Permutation>
         return permutation.Length;
     }
 
-    public int[] GetPermutation()
+    public long[] GetPermutation()
     {
         return permutation;
     }
@@ -114,7 +114,7 @@ public class Permutation : IComparable<Permutation>
         return 0;
     }
 
-    public bool Compare(int[] permutation)
+    public bool Compare(long[] permutation)
     {
         return this.permutation.SequenceEqual(permutation);
     }
